==> 3-PD Model - Estimation.ipynb <==
import pandas as pd
pd.set_option("display.max_columns", None)
pd.set_option("display.max_rows", None)
import numpy as np

loan_data_inputs_train = pd.read_csv(r'D:\Datasets\LendingClub\loan_data_inputs_train.csv', index_col=0)
loan_data_targets_train = pd.read_csv(r'D:\Datasets\LendingClub\loan_data_targets_train.csv', index_col=0).values.reshape(-1,)
loan_data_inputs_test = pd.read_csv(r'D:\Datasets\LendingClub\loan_data_inputs_test.csv', index_col=0)
loan_data_targets_test = pd.read_csv(r'D:\Datasets\LendingClub\loan_data_targets_test.csv', index_col=0).values.reshape(-1,)

print(loan_data_inputs_train.shape)
print(loan_data_targets_train.shape)
print(loan_data_inputs_test.shape)
print(loan_data_targets_test.shape)

# logistic regression
# easy to understand and interpret
# logistic function
# P(X) = e**X / (1 - e**X)
# probability of event is exponential of linear combination of coefficients dividend by 1 + the same exponential
# P(of event occuring)/1-P(of event not occurring) = exponential of linear combination of coefficients
# log odds
# dice
# 1/6 (probability of getting a 1)
# 5/6 (probability of not getting a 1)
# 1:5 odds or 0.2
# log odds (P(Y=1)/1-P(Y=0)) = B0 + B1X1 + B2X2+ ... BmXm
# all the independent variables either 0 or 1
# the coefficient shows us the difference in log odds caused in the final estimation by changing the variable from 0 to 1
# to get the odds we take e**(coefficient[B1])
# eg. if e**(B1) = 2.25
# and B1 is higher having a mortgage and not being in default
# 2.25 times the odds of having a mortgage and being in default
# 125% more likely to not be in default if have a mortgage than not

### Selecting the Features

inputs_train_with_ref_cat = loan_data_inputs_train.loc[: , ['grade:A',
'grade:B',
'grade:C',
'grade:D',
'grade:E',
'grade:F',
'grade:G',
'home_ownership:RENT_OTHER_NONE_ANY',
'home_ownership:OWN',
'home_ownership:MORTGAGE',
'addr_state:ME_DC_WY_WV_VT_MS',
'addr_state:NY',
'addr_state:RI_NM_NC_OK_SD_VA_LA',
'addr_state:MD_NJ_MO',
'addr_state:CA',
'addr_state:ID_TN',
'addr_state:AZ_MI_UT_AR_PA_OH_MN_IN',
'addr_state:KY_MA_GA_WA_WI_DE',
'addr_state:OR_IL',
'addr_state:TX',
'addr_state:CT_SC_AK',
'addr_state:KS_CO_MT_NH',
'verification_status:Verified',
'verification_status:Source Verified',
'verification_status:Not Verified',
'purpose:edu_sb_re_m_h_other',
'purpose:med_wed_vac',
'purpose:debt_consolidation',
'purpose:hi_mp_car',
'purpose:credit_card',
'initial_list_status:f',
'initial_list_status:w',
'term:36',
'term:60',
'emp_length:0',
'emp_length:1',
'emp_length:2-4',
'emp_length:5-6',
'emp_length:7-9',
'emp_length:10',
'mths_since_issue_d:<38',
'mths_since_issue_d:38-39',
'mths_since_issue_d:<40-41',
'mths_since_issue_d:<42-48',
'mths_since_issue_d:<49-52',
'mths_since_issue_d:<53-64',
'mths_since_issue_d:<65-84',
'mths_since_issue_d:>84',
'int_rate:<9.548',
'int_rate:9.548-12.025',
'int_rate:12.025-15.74',
'int_rate:15.74-20.281',
'int_rate:>20.281',
'mths_since_earliest_cr_line:<142',
'mths_since_earliest_cr_line:142-266',
'mths_since_earliest_cr_line:266-355',
'mths_since_earliest_cr_line:>355',
'delinq_2yrs:0',
'delinq_2yrs:1-3',
'delinq_2yrs:>=4',
'inq_last_6mths:0',
'inq_last_6mths:1-2',
'inq_last_6mths:3-6',
'inq_last_6mths:>6',
'open_acc:0',
'open_acc:1-3',
'open_acc:4-12',
'open_acc:13-17',
'open_acc:18-22',
'open_acc:23-25',
'open_acc:26-30',
'open_acc:>=31',
'pub_rec:0-2',
'pub_rec:3-4',
'pub_rec:>=5',
'total_acc:<=25',
'total_acc:25-51',
'total_acc:>=51',
'acc_now_delinq:0',
'acc_now_delinq:>=1',
'total_rev_hi_lim:<=5K',
'total_rev_hi_lim:5K-10K',
'total_rev_hi_lim:10K-20K',
'total_rev_hi_lim:20K-30K',
'total_rev_hi_lim:30K-40K',
'total_rev_hi_lim:40K-55K',
'total_rev_hi_lim:55K-95K',
'total_rev_hi_lim:>95K',
'annual_inc:<20K',
'annual_inc:20K-30K',
'annual_inc:30K-40K',
'annual_inc:40K-50K',
'annual_inc:50K-60K',
'annual_inc:60K-70K',
'annual_inc:70K-80K',
'annual_inc:80K-90K',
'annual_inc:90K-100K',
'annual_inc:100K-120K',
'annual_inc:120K-140K',
'annual_inc:>140K',
'mths_since_last_delinq:Missing',
'mths_since_last_delinq:0-3',
'mths_since_last_delinq:4-30',
'mths_since_last_delinq:31-56',
'mths_since_last_delinq:>=57',
'dti:<=1.4',
'dti:1.4-3.5',
'dti:3.5-7.7',
'dti:7.7-10.5',
'dti:10.5-16.1',
'dti:16.1-20.3',
'dti:20.3-21.7',
'dti:21.7-22.4',
'dti:22.4-35',
'dti:>35',
'mths_since_last_record:Missing',
'mths_since_last_record:0-2',
'mths_since_last_record:3-20',
'mths_since_last_record:21-31',
'mths_since_last_record:32-80',
'mths_since_last_record:81-86',
'mths_since_last_record:>=86',
]]

# removing one dummy variable for k-1
ref_categories = ['grade:G',
                'home_ownership:RENT_OTHER_NONE_ANY',
                'addr_state:ND_NE_IA_NV_HI_FL_AL',
                'verification_status:Verified',
                'purpose:edu_sb_re_m_h_other',
                 'initial_list_status:f',
                 'term:60',
                 'emp_length:0',
                 'mths_since_issue_d:>84',
                 'int_rate:>20.281',
                 'mths_since_earliest_cr_line:<142',
                 'delinq_2yrs:>=4',
                 'inq_last_6mths:>6',
                 'open_acc:0',
                 'pub_rec:>=5',
                 'total_acc:<=25',
                 'acc_now_delinq:>=1',
                 'total_rev_hi_lim<=5K',
                 'annual_inc:<20K',
                 'mths_since_last_delinq:Missing',
                 'dti>35',
                 'mths_since_last_record:Missing']

# selecting the desires columns
sel_cols = [c for c in inputs_train_with_ref_cat.columns if c not in ref_categories]

# selecting columns that are not reference categories
inputs_train = loan_data_inputs_train[sel_cols]
# loan_data_targets_train
inputs_test = loan_data_inputs_test[inputs_train.columns]
# loan_data_targets_test

from sklearn.linear_model import LogisticRegression

reg = LogisticRegression(max_iter=2000)

reg.fit(inputs_train, loan_data_targets_train)

reg.intercept_

reg.coef_

feature_name = inputs_train.columns.values

summary_table = pd.DataFrame(columns = ['Feature name'], data = feature_name)
summary_table['Coefficients'] = np.transpose(reg.coef_)
# add the intercept
summary_table.index = summary_table.index + 1
summary_table.loc[0] = ['Intercept', reg.intercept_[0]]
summary_table = summary_table.sort_index()
summary_table

### Build a logistic modele with multivariate p-values

# P values for sklearn logistic regression.

# Class to display p-values for logistic regression in sklearn.

from sklearn import linear_model
import scipy.stats as stat

class LogisticRegression_with_p_values:
    
    def __init__(self,*args,**kwargs):#,**kwargs):
        self.model = linear_model.LogisticRegression(*args,**kwargs)#,**args)

    def fit(self,X,y):
        self.model.fit(X,y)
        
        #### Get p-values for the fitted model ####
        denom = (2.0 * (1.0 + np.cosh(self.model.decision_function(X))))
        denom = np.tile(denom,(X.shape[1],1)).T
        F_ij = np.dot((X / denom).T,X) ## Fisher Information Matrix
        Cramer_Rao = np.linalg.inv(F_ij) ## Inverse Information Matrix
        sigma_estimates = np.sqrt(np.diagonal(Cramer_Rao))
        z_scores = self.model.coef_[0] / sigma_estimates # z-score for eaach model coefficient
        p_values = [stat.norm.sf(abs(x)) * 2 for x in z_scores] ### two tailed test for p-values
        
        self.coef_ = self.model.coef_
        self.intercept_ = self.model.intercept_
        self.p_values = p_values

reg = LogisticRegression_with_p_values(max_iter=2000)

reg.fit(inputs_train, loan_data_targets_train)

summary_table = pd.DataFrame(columns = ['Feature name'], data = feature_name)
summary_table['Coefficients'] = np.transpose(reg.coef_)
# add the intercept
summary_table.index = summary_table.index + 1
summary_table.loc[0] = ['Intercept', reg.intercept_[0]]
summary_table = summary_table.sort_index()
summary_table

p_values = reg.p_values
# added nan for intercept
p_values = np.append(np.nan, np.array(p_values))

summary_table['p_values'] = p_values

summary_table

# if one or few dummy variables representing one original indendent variable are
# statistically significant, it would be best to retain all the variables related that original independent variable
# if p < 0.05 = statistically significant
# remove 
#
# delinq_2yrs
# open_acc
# pub_rec
# total_rev_hi_lim
# total_acc
# 
# as not statistically significant

inputs_train_with_ref_cat = loan_data_inputs_train.loc[: , ['grade:A',
'grade:B',
'grade:C',
'grade:D',
'grade:E',
'grade:F',
'grade:G',
'home_ownership:RENT_OTHER_NONE_ANY',
'home_ownership:OWN',
'home_ownership:MORTGAGE',
'addr_state:ME_DC_WY_WV_VT_MS',
'addr_state:NY',
'addr_state:RI_NM_NC_OK_SD_VA_LA',
'addr_state:MD_NJ_MO',
'addr_state:CA',
'addr_state:ID_TN',
'addr_state:AZ_MI_UT_AR_PA_OH_MN_IN',
'addr_state:KY_MA_GA_WA_WI_DE',
'addr_state:OR_IL',
'addr_state:TX',
'addr_state:CT_SC_AK',
'addr_state:KS_CO_MT_NH',
'verification_status:Verified',
'verification_status:Source Verified',
'verification_status:Not Verified',
'purpose:edu_sb_re_m_h_other',
'purpose:med_wed_vac',
'purpose:debt_consolidation',
'purpose:hi_mp_car',
'purpose:credit_card',
'initial_list_status:f',
'initial_list_status:w',
'term:36',
'term:60',
'emp_length:0',
'emp_length:1',
'emp_length:2-4',
'emp_length:5-6',
'emp_length:7-9',
'emp_length:10',
'mths_since_issue_d:<38',
'mths_since_issue_d:38-39',
'mths_since_issue_d:<40-41',
'mths_since_issue_d:<42-48',
'mths_since_issue_d:<49-52',
'mths_since_issue_d:<53-64',
'mths_since_issue_d:<65-84',
'mths_since_issue_d:>84',
'int_rate:<9.548',
'int_rate:9.548-12.025',
'int_rate:12.025-15.74',
'int_rate:15.74-20.281',
'int_rate:>20.281',
'mths_since_earliest_cr_line:<142',
'mths_since_earliest_cr_line:142-266',
'mths_since_earliest_cr_line:266-355',
'mths_since_earliest_cr_line:>355',
#'delinq_2yrs:0',
#'delinq_2yrs:1-3',
#'delinq_2yrs:>=4',
'inq_last_6mths:0',
'inq_last_6mths:1-2',
'inq_last_6mths:3-6',
'inq_last_6mths:>6',
#'open_acc:0',
#'open_acc:1-3',
#'open_acc:4-12',
#'open_acc:13-17',
#'open_acc:18-22',
#'open_acc:23-25',
#'open_acc:26-30',
#'open_acc:>=31',
#'pub_rec:0-2',
#'pub_rec:3-4',
#'pub_rec:>=5',
#'total_acc:<=25',
#'total_acc:25-51',
#'total_acc:>=51',
'acc_now_delinq:0',
'acc_now_delinq:>=1',
#'total_rev_hi_lim:<=5K',
#'total_rev_hi_lim:5K-10K',
#'total_rev_hi_lim:10K-20K',
#'total_rev_hi_lim:20K-30K',
#'total_rev_hi_lim:30K-40K',
#'total_rev_hi_lim:40K-55K',
#'total_rev_hi_lim:55K-95K',
#'total_rev_hi_lim:>95K',
'annual_inc:<20K',
'annual_inc:20K-30K',
'annual_inc:30K-40K',
'annual_inc:40K-50K',
'annual_inc:50K-60K',
'annual_inc:60K-70K',
'annual_inc:70K-80K',
'annual_inc:80K-90K',
'annual_inc:90K-100K',
'annual_inc:100K-120K',
'annual_inc:120K-140K',
'annual_inc:>140K',
'mths_since_last_delinq:Missing',
'mths_since_last_delinq:0-3',
'mths_since_last_delinq:4-30',
'mths_since_last_delinq:31-56',
'mths_since_last_delinq:>=57',
'dti:<=1.4',
'dti:1.4-3.5',
'dti:3.5-7.7',
'dti:7.7-10.5',
'dti:10.5-16.1',
'dti:16.1-20.3',
'dti:20.3-21.7',
'dti:21.7-22.4',
'dti:22.4-35',
'dti:>35',
'mths_since_last_record:Missing',
'mths_since_last_record:0-2',
'mths_since_last_record:3-20',
'mths_since_last_record:21-31',
'mths_since_last_record:32-80',
'mths_since_last_record:81-86',
'mths_since_last_record:>=86',
]]

# selecting the desires columns
sel_cols = [c for c in inputs_train_with_ref_cat.columns if c not in ref_categories]

# selecting columns that are not reference categories
inputs_train = loan_data_inputs_train[sel_cols]
# loan_data_targets_train
inputs_test = loan_data_inputs_test[inputs_train.columns]
# loan_data_targets_test

reg2 = LogisticRegression_with_p_values(max_iter=2000)
reg2.fit(inputs_train, loan_data_targets_train)

feature_name = inputs_train.columns.values

summary_table = pd.DataFrame(columns = ['Feature name'], data = feature_name)
summary_table['Coefficients'] = np.transpose(reg2.coef_)
# add the intercept
summary_table.index = summary_table.index + 1
summary_table.loc[0] = ['Intercept', reg2.intercept_[0]]
summary_table = summary_table.sort_index()
p_values = reg2.p_values
# added nan for intercept
p_values = np.append(np.nan, np.array(p_values))
summary_table['p_values'] = p_values
summary_table

### Saving the model

import pickle

pickle.dump(reg2, open('pd_model.sav', 'wb'))

## Interpreting coefficients

# base of natural logarithm to the exponent of the coefficient gives us the ratio of odds of default occurring given that coefficient being present to that coefficient not being present
# higher coefficients = greater odds of being a good borrower, or lower odds of default
# odds are always relative to the reference category
# AAA rating has coefficient 1.2
# CCC reference category 
# means AAA rating e**1.2 = 3.3 times greater odds of being a non-defaulting borrower than the reference category
# if BBB rating has coefficient 0.8 means e**0.8 = 2.2 times greater odds of being non-default than reference category
# can use this relationship to compare odds between categories
# 3.3/2.2 = 1.5, AAA 1.5 times greater odds of being non-defaulting borrower than BBB
# odds between categories is e**(B0 - B1) exponent to difference of coefficients

inputs_train.to_csv(r'D:\Datasets\LendingClub\inputs_train_with_ref_cat.csv')
inputs_test.to_csv(r'D:\Datasets\LendingClub\inputs_test_with_ref_cat.csv')
