==> 6-PD Model - Monitoring - New Data Preparation.ipynb <==
# population stability index (PSI)
# new data
# present data (actual) and past data (expected) distributions of features
# 
# PSI = sigma[(%actualj - %expectedj) x ln(%actualj/%expectedj)]
# 0 and 1
# 0 = no difference
# 1 = absolute difference
# < 0.1 = little to no difference
# 0.1 < PSI < 0.25 = little difference (no action taken)
# PSI > 0.25 = big difference (action taken)

import numpy as np
import pandas as pd
pd.options.display.max_columns = None

loan_data_inputs_train = pd.read_csv('D:\Datasets\LendingClub\loan_data_inputs_train.csv', index_col=0)
loan_data_targets_train = pd.read_csv('D:\Datasets\LendingClub\loan_data_targets_train.csv', index_col=0)
loan_data_inputs_test = pd.read_csv('D:\Datasets\LendingClub\loan_data_inputs_test.csv', index_col=0)
loan_data_targets_test = pd.read_csv('D:\Datasets\LendingClub\loan_data_targets_test.csv', index_col=0)

loan_data_backup = pd.read_csv(r'D:\Datasets\LendingClub\loan_data_2015.csv')
loan_data_backup.head()

loan_data_backup.info()

loan_data = loan_data_backup.copy()

loan_data['term_int'] = loan_data['term'].str.replace(' ', '')
loan_data['term_int'] = loan_data['term_int'].str.replace('months', '')
loan_data['term_int'] = pd.to_numeric(loan_data['term_int'])

loan_data['term:36'] = np.where((loan_data['term_int']==36), 1, 0)
loan_data['term:60'] = np.where((loan_data['term_int']==60), 1, 0)

loan_data['emp_length_int'] = loan_data['emp_length'].str.replace('\+ years', '')
loan_data['emp_length_int'] = loan_data['emp_length_int'].str.replace('< 1 year', str(0))
loan_data['emp_length_int'] = loan_data['emp_length_int'].str.replace('n/a', str(0))
loan_data['emp_length_int'] = loan_data['emp_length_int'].str.replace(' years', '')
loan_data['emp_length_int'] = loan_data['emp_length_int'].str.replace(' year', '')

loan_data['empy_length_int'] = pd.to_numeric(loan_data['emp_length_int'])

loan_data['term_int'] = loan_data['term'].str.replace(' ', '')
loan_data['term_int'] = loan_data['term_int'].str.replace('months', '')
loan_data['term_int'] = pd.to_numeric(loan_data['term_int'])

loan_data['earliest_cr_line_date'] = pd.to_datetime(loan_data['earliest_cr_line'], format='%b-%y')

# Fixing incorrectly imputed dates
loan_data['earliest_cr_line_date'] = [pd.to_datetime(str(e).replace('20', '19')) if e > pd.to_datetime('2017-12-01') else e for e in loan_data['earliest_cr_line_date']]

loan_data['mths_since_earliest_cr_line'] = round(pd.to_numeric((pd.to_datetime('2017-12-01') - loan_data['earliest_cr_line_date']) / np.timedelta64(1, 'M')))

# Create 'mths_since_issue_d' with months since issue date
loan_data['mths_since_issue_d'] = pd.to_datetime(loan_data['issue_d'], format='%b-%y')
loan_data['mths_since_issue_d'] = round(pd.to_numeric((pd.to_datetime('2017-12-01') - loan_data['mths_since_issue_d']) / np.timedelta64(1, 'M')))

loan_data_dummies = [pd.get_dummies(loan_data['grade'], prefix='grade', prefix_sep=':'),
                    pd.get_dummies(loan_data['sub_grade'], prefix='sub_grade', prefix_sep=':'),
                    pd.get_dummies(loan_data['home_ownership'], prefix='home_ownership', prefix_sep=':'),
                    pd.get_dummies(loan_data['verification_status'], prefix='verification_status', prefix_sep=':'),
                    pd.get_dummies(loan_data['loan_status'], prefix='loan_status', prefix_sep=':'),
                    pd.get_dummies(loan_data['purpose'], prefix='purpose', prefix_sep=':'),
                    pd.get_dummies(loan_data['addr_state'], prefix='addr_state', prefix_sep=':'),
                    pd.get_dummies(loan_data['initial_list_status'], prefix='initial_list_status', prefix_sep=':')]

loan_data_dummies = pd.concat(loan_data_dummies, axis=1)

loan_data = pd.concat([loan_data, loan_data_dummies], axis=1)

# total_rev_hi_lim = total revolving limit
loan_data['total_rev_hi_lim'].fillna(loan_data['funded_amnt'], inplace=True)
loan_data['total_rev_hi_lim'].isnull().sum()

loan_data['annual_inc'].fillna((loan_data['annual_inc'].dropna()).mean(), inplace=True)
loan_data['annual_inc'].isnull().sum()

for c in ['mths_since_earliest_cr_line', 'acc_now_delinq', 'total_acc', 'pub_rec', 'open_acc', 'inq_last_6mths', 'delinq_2yrs', 'emp_length_int']:
    loan_data[c].fillna(0, inplace=True)
    print(c + ' : ' + str(loan_data[c].isnull().sum()))

loan_data.to_csv(r'D:\Datasets\LendingClub\loan_data_processed_2015.csv')

loan_data.columns.values

# positive coefficients = more creditworthiness
# negative coefficients = less creditworthiness

# np.where(condition, value if true, value if false)
loan_data['good_bad'] = np.where(loan_data['loan_status'].isin(['Charged Off',
                                                               'Defeault',
                                                               'Does not meet the credit policy. Status:Charged Off',
                                                               'Late (31-120 days)']), 0, 1)

#from sklearn.model_selection import train_test_split

#loan_data_inputs_train, loan_data_inputs_test, loan_data_targets_train, loan_data_targets_test = train_test_split(loan_data.drop('good_bad', axis=1), loan_data['good_bad'], test_size=0.2, random_state=1)

# the below are uncommented to run the processing through the test data
#df_inputs_prepr = loan_data_inputs_train
#df_targets_prepr = loan_data_targets_train
#df_inputs_prepr = loan_data_inputs_test
#df_targets_prepr = loan_data_targets_test
df_inputs_prepr = loan_data.drop('good_bad', axis=1)
df_targets_prepr = loan_data['good_bad']

df1 = pd.concat([df_inputs_prepr['grade'], df_targets_prepr], axis=1)

# OTHER, NONE and ANY have very little observations, so should not create dummies for these variables
# but we can combine these underrepresented categories into a new category together with the RENT variable
# OWN and MORTGAGE will be seperate dummy variables


df_inputs_prepr['home_ownership:RENT_OTHER_NONE_ANY'] = sum([df_inputs_prepr['home_ownership:RENT'], df_inputs_prepr['home_ownership:ANY']])

if ['addr_state:ND'] in df_inputs_prepr.columns.values:
    pass
else:
    df_inputs_prepr['addr_state:ND'] = 0
    
if ['addr_state:IA'] in df_inputs_prepr.columns.values:
    pass
else:
    df_inputs_prepr['addr_state:IA'] = 0
    

if ['addr_state:ID'] in df_inputs_prepr.columns.values:
    pass
else:
    df_inputs_prepr['addr_state:ID'] = 0

df_inputs_prepr['addr_state:ND_NE_IA_NV_HI_FL_AL'] = sum([df_inputs_prepr['addr_state:ND'], df_inputs_prepr['addr_state:NE'],
                                                        df_inputs_prepr['addr_state:IA'], df_inputs_prepr['addr_state:NV'],
                                                        df_inputs_prepr['addr_state:HI'], df_inputs_prepr['addr_state:FL'],
                                                        df_inputs_prepr['addr_state:AL']])

df_inputs_prepr['addr_state:ME_DC_WY_WV_VT_MS'] = sum([df_inputs_prepr['addr_state:ME'], df_inputs_prepr['addr_state:DC'],
                                                        df_inputs_prepr['addr_state:WY'], df_inputs_prepr['addr_state:WV'],
                                                        df_inputs_prepr['addr_state:VT'], df_inputs_prepr['addr_state:MS']])

df_inputs_prepr['addr_state:RI_NM_NC_OK_SD_VA_LA'] = sum([df_inputs_prepr['addr_state:RI'], df_inputs_prepr['addr_state:NM'],
                                                        df_inputs_prepr['addr_state:NC'], df_inputs_prepr['addr_state:OK'],
                                                        df_inputs_prepr['addr_state:SD'], df_inputs_prepr['addr_state:VA'],
                                                        df_inputs_prepr['addr_state:LA']])

df_inputs_prepr['addr_state:MD_NJ_MO'] = sum([df_inputs_prepr['addr_state:MD'], df_inputs_prepr['addr_state:NJ'], df_inputs_prepr['addr_state:MO']])

df_inputs_prepr['addr_state:ID_TN'] = sum([df_inputs_prepr['addr_state:ID'], df_inputs_prepr['addr_state:TN']])

df_inputs_prepr['addr_state:AZ_MI_UT_AR_PA_OH_MN_IN'] = sum([df_inputs_prepr['addr_state:AZ'], df_inputs_prepr['addr_state:MI'],
                                                        df_inputs_prepr['addr_state:UT'], df_inputs_prepr['addr_state:AR'],
                                                        df_inputs_prepr['addr_state:PA'], df_inputs_prepr['addr_state:OH'],
                                                        df_inputs_prepr['addr_state:MN'], df_inputs_prepr['addr_state:IN']])

df_inputs_prepr['addr_state:KY_MA_GA_WA_WI_DE'] = sum([df_inputs_prepr['addr_state:KY'], df_inputs_prepr['addr_state:MA'],
                                                        df_inputs_prepr['addr_state:GA'], df_inputs_prepr['addr_state:WA'],
                                                        df_inputs_prepr['addr_state:WI'], df_inputs_prepr['addr_state:DE']])

df_inputs_prepr['addr_state:OR_IL'] = sum([df_inputs_prepr['addr_state:OR'], df_inputs_prepr['addr_state:IL']])

df_inputs_prepr['addr_state:CT_SC_AK'] = sum([df_inputs_prepr['addr_state:CT'], df_inputs_prepr['addr_state:SC'],
                                                        df_inputs_prepr['addr_state:AK']])

df_inputs_prepr['addr_state:KS_CO_MT_NH'] = sum([df_inputs_prepr['addr_state:KS'], df_inputs_prepr['addr_state:CO'],
                                                        df_inputs_prepr['addr_state:MT'], df_inputs_prepr['addr_state:NH']])

# we put educational, small_business, renewable_energy, moving, house and other in one category (this will be the reference category)
# we put medical, wedding, vacation in one category
# we put debt_consolidation in its one category
# we put home_improvement, major_purchase and car in one category
# we put credit_card in one category

df_inputs_prepr['purpose:edu_sb_re_m_h_other'] = sum([df_inputs_prepr['purpose:educational'], df_inputs_prepr['purpose:small_business'],
                                                        df_inputs_prepr['purpose:renewable_energy'], df_inputs_prepr['purpose:moving'],
                                                        df_inputs_prepr['purpose:house'], df_inputs_prepr['purpose:other']])

df_inputs_prepr['purpose:med_wed_vac'] = sum([df_inputs_prepr['purpose:medical'], df_inputs_prepr['purpose:wedding'],
                                                        df_inputs_prepr['purpose:vacation']])

df_inputs_prepr['purpose:hi_mp_car'] = sum([df_inputs_prepr['purpose:home_improvement'], df_inputs_prepr['purpose:major_purchase'],
                                                        df_inputs_prepr['purpose:car']])

# 0 seperate (reference category)
# 1 seperate
# 2-4 together
# 5-6 together
# 7-9 together
# 10 seperate

df_inputs_prepr['emp_length:0'] = np.where(df_inputs_prepr['emp_length_int'].isin([0]), 1, 0)
df_inputs_prepr['emp_length:1'] = np.where(df_inputs_prepr['emp_length_int'].isin([1]), 1, 0)
df_inputs_prepr['emp_length:2-4'] = np.where(df_inputs_prepr['emp_length_int'].isin(range(2, 5)), 1, 0)
df_inputs_prepr['emp_length:5-6'] = np.where(df_inputs_prepr['emp_length_int'].isin(range(5, 7)), 1, 0)
df_inputs_prepr['emp_length:7-9'] = np.where(df_inputs_prepr['emp_length_int'].isin(range(7, 10)), 1, 0)
df_inputs_prepr['emp_length:10'] = np.where(df_inputs_prepr['emp_length_int'].isin([10]), 1, 0)

df_inputs_prepr['mths_since_issue_d:<38'] = np.where(df_inputs_prepr['mths_since_issue_d'].isin(range(38)), 1, 0)
df_inputs_prepr['mths_since_issue_d:38-39'] = np.where(df_inputs_prepr['mths_since_issue_d'].isin(range(38, 40)), 1, 0)
df_inputs_prepr['mths_since_issue_d:<40-41'] = np.where(df_inputs_prepr['mths_since_issue_d'].isin(range(40, 42)), 1, 0)
df_inputs_prepr['mths_since_issue_d:<40-41'] = np.where(df_inputs_prepr['mths_since_issue_d'].isin(range(40, 42)), 1, 0)
df_inputs_prepr['mths_since_issue_d:<42-48'] = np.where(df_inputs_prepr['mths_since_issue_d'].isin(range(42, 49)), 1, 0)
df_inputs_prepr['mths_since_issue_d:<49-52'] = np.where(df_inputs_prepr['mths_since_issue_d'].isin(range(49, 53)), 1, 0)
df_inputs_prepr['mths_since_issue_d:<53-64'] = np.where(df_inputs_prepr['mths_since_issue_d'].isin(range(53, 65)), 1, 0)
df_inputs_prepr['mths_since_issue_d:<65-84'] = np.where(df_inputs_prepr['mths_since_issue_d'].isin(range(65, 85)), 1, 0)
df_inputs_prepr['mths_since_issue_d:>84'] = np.where(df_inputs_prepr['mths_since_issue_d'].isin(range(85, int(df_inputs_prepr['mths_since_issue_d'].max()))), 1, 0)

# 5.4 - 9.548 in one category
# 9.548 - 12.025 in one category
# 12.025 - 15.74 in one category
# 15.74 - 20.281
# > 20.281
df_inputs_prepr['int_rate:<9.548'] = np.where((df_inputs_prepr['int_rate'] <= 9.548), 1, 0)
df_inputs_prepr['int_rate:9.548-12.025'] = np.where((df_inputs_prepr['int_rate'] > 9.548) & (df_inputs_prepr['int_rate'] <= 12.025), 1, 0)
df_inputs_prepr['int_rate:12.025-15.74'] = np.where((df_inputs_prepr['int_rate'] > 12.025) & (df_inputs_prepr['int_rate'] <= 15.74), 1, 0)
df_inputs_prepr['int_rate:15.74-20.281'] = np.where((df_inputs_prepr['int_rate'] > 15.74) & (df_inputs_prepr['int_rate'] <= 20.281), 1, 0)
df_inputs_prepr['int_rate:>20.281'] = np.where((df_inputs_prepr['int_rate'] > 20.281), 1, 0)

# < 141.92
# 141.92 - 266.1
# 266.1 - 354.8
# > 354.8

df_inputs_prepr['mths_since_earliest_cr_line:<142'] = np.where(df_inputs_prepr['mths_since_earliest_cr_line'].isin(range(142)), 1, 0)
df_inputs_prepr['mths_since_earliest_cr_line:142-266'] = np.where(df_inputs_prepr['mths_since_earliest_cr_line'].isin(range(142, 266)), 1, 0)
df_inputs_prepr['mths_since_earliest_cr_line:266-355'] = np.where(df_inputs_prepr['mths_since_earliest_cr_line'].isin(range(266, 355)), 1, 0)
df_inputs_prepr['mths_since_earliest_cr_line:>355'] = np.where(df_inputs_prepr['mths_since_earliest_cr_line'].isin(range(355, int(df_inputs_prepr['mths_since_earliest_cr_line'].max()))), 1, 0)

df_inputs_prepr['delinq_2yrs:0'] = np.where((df_inputs_prepr['delinq_2yrs'] == 0), 1, 0)
df_inputs_prepr['delinq_2yrs:1-3'] = np.where((df_inputs_prepr['delinq_2yrs'] >= 1) & (df_inputs_prepr['delinq_2yrs'] <= 3), 1, 0)
df_inputs_prepr['delinq_2yrs:>=4'] = np.where((df_inputs_prepr['delinq_2yrs'] >= 4), 1, 0)

df_inputs_prepr['inq_last_6mths:0'] = np.where((df_inputs_prepr['inq_last_6mths'] == 0), 1, 0)
df_inputs_prepr['inq_last_6mths:1-2'] = np.where((df_inputs_prepr['inq_last_6mths'] >= 1) & (df_inputs_prepr['inq_last_6mths'] <= 2), 1, 0)
df_inputs_prepr['inq_last_6mths:3-6'] = np.where((df_inputs_prepr['inq_last_6mths'] >= 3) & (df_inputs_prepr['inq_last_6mths'] <= 6), 1, 0)
df_inputs_prepr['inq_last_6mths:>6'] = np.where((df_inputs_prepr['inq_last_6mths'] > 6), 1, 0)

df_inputs_prepr['open_acc:0'] = np.where((df_inputs_prepr['open_acc'] == 0), 1, 0)
df_inputs_prepr['open_acc:1-3'] = np.where((df_inputs_prepr['open_acc'] >= 1) & (df_inputs_prepr['open_acc'] <= 3), 1, 0)
df_inputs_prepr['open_acc:4-12'] = np.where((df_inputs_prepr['open_acc'] >= 4) & (df_inputs_prepr['open_acc'] <= 12), 1, 0)
df_inputs_prepr['open_acc:13-17'] = np.where((df_inputs_prepr['open_acc'] >= 13) & (df_inputs_prepr['open_acc'] <= 17), 1, 0)
df_inputs_prepr['open_acc:18-22'] = np.where((df_inputs_prepr['open_acc'] >= 18) & (df_inputs_prepr['open_acc'] <= 22), 1, 0)
df_inputs_prepr['open_acc:23-25'] = np.where((df_inputs_prepr['open_acc'] >= 23) & (df_inputs_prepr['open_acc'] <= 25), 1, 0)
df_inputs_prepr['open_acc:26-30'] = np.where((df_inputs_prepr['open_acc'] >= 26) & (df_inputs_prepr['open_acc'] <= 30), 1, 0)
df_inputs_prepr['open_acc:>=31'] = np.where((df_inputs_prepr['open_acc'] >= 31), 1, 0)

df_inputs_prepr['pub_rec:0-2'] = np.where((df_inputs_prepr['pub_rec'] >= 0) & (df_inputs_prepr['pub_rec'] <= 2), 1, 0)
df_inputs_prepr['pub_rec:3-4'] = np.where((df_inputs_prepr['pub_rec'] >= 3) & (df_inputs_prepr['pub_rec'] <= 4), 1, 0)
df_inputs_prepr['pub_rec:>=5'] = np.where((df_inputs_prepr['pub_rec'] >= 5), 1, 0)

df_inputs_prepr['total_acc:<=25'] = np.where((df_inputs_prepr['total_acc'] <= 25), 1, 0)
df_inputs_prepr['total_acc:25-51'] = np.where((df_inputs_prepr['total_acc'] >= 25) & (df_inputs_prepr['total_acc'] <= 51), 1, 0)
df_inputs_prepr['total_acc:>=51'] = np.where((df_inputs_prepr['total_acc'] >= 51), 1, 0)

df_inputs_prepr['acc_now_delinq:0'] = np.where((df_inputs_prepr['acc_now_delinq'] == 0), 1, 0)
df_inputs_prepr['acc_now_delinq:>=1'] = np.where((df_inputs_prepr['acc_now_delinq'] >= 1), 1, 0)

df_inputs_prepr['total_rev_hi_lim:<=5K'] = np.where((df_inputs_prepr['total_rev_hi_lim'] <= 5000), 1, 0)
df_inputs_prepr['total_rev_hi_lim:5K-10K'] = np.where((df_inputs_prepr['total_rev_hi_lim'] > 5000) & (df_inputs_prepr['total_rev_hi_lim'] <= 10000), 1, 0)
df_inputs_prepr['total_rev_hi_lim:10K-20K'] = np.where((df_inputs_prepr['total_rev_hi_lim'] > 10000) & (df_inputs_prepr['total_rev_hi_lim'] <= 20000), 1, 0)
df_inputs_prepr['total_rev_hi_lim:20K-30K'] = np.where((df_inputs_prepr['total_rev_hi_lim'] > 20000) & (df_inputs_prepr['total_rev_hi_lim'] <= 30000), 1, 0)
df_inputs_prepr['total_rev_hi_lim:30K-40K'] = np.where((df_inputs_prepr['total_rev_hi_lim'] > 30000) & (df_inputs_prepr['total_rev_hi_lim'] <= 40000), 1, 0)
df_inputs_prepr['total_rev_hi_lim:40K-55K'] = np.where((df_inputs_prepr['total_rev_hi_lim'] > 40000) & (df_inputs_prepr['total_rev_hi_lim'] <= 55000), 1, 0)
df_inputs_prepr['total_rev_hi_lim:55K-95K'] = np.where((df_inputs_prepr['total_rev_hi_lim'] > 55000) & (df_inputs_prepr['total_rev_hi_lim'] <= 95000), 1, 0)
df_inputs_prepr['total_rev_hi_lim:>95K'] = np.where((df_inputs_prepr['total_rev_hi_lim'] > 95000), 1, 0)

df_inputs_prepr['annual_inc:<20K'] = np.where((df_inputs_prepr['annual_inc'] <= 20000), 1, 0)
df_inputs_prepr['annual_inc:20K-30K'] = np.where((df_inputs_prepr['annual_inc'] > 20000) & (df_inputs_prepr['annual_inc'] <= 30000), 1, 0)
df_inputs_prepr['annual_inc:30K-40K'] = np.where((df_inputs_prepr['annual_inc'] > 30000) & (df_inputs_prepr['annual_inc'] <= 40000), 1, 0)
df_inputs_prepr['annual_inc:40K-50K'] = np.where((df_inputs_prepr['annual_inc'] > 40000) & (df_inputs_prepr['annual_inc'] <= 50000), 1, 0)
df_inputs_prepr['annual_inc:50K-60K'] = np.where((df_inputs_prepr['annual_inc'] > 50000) & (df_inputs_prepr['annual_inc'] <= 60000), 1, 0)
df_inputs_prepr['annual_inc:60K-70K'] = np.where((df_inputs_prepr['annual_inc'] > 60000) & (df_inputs_prepr['annual_inc'] <= 70000), 1, 0)
df_inputs_prepr['annual_inc:70K-80K'] = np.where((df_inputs_prepr['annual_inc'] > 70000) & (df_inputs_prepr['annual_inc'] <= 80000), 1, 0)
df_inputs_prepr['annual_inc:80K-90K'] = np.where((df_inputs_prepr['annual_inc'] > 80000) & (df_inputs_prepr['annual_inc'] <= 90000), 1, 0)
df_inputs_prepr['annual_inc:90K-100K'] = np.where((df_inputs_prepr['annual_inc'] > 90000) & (df_inputs_prepr['annual_inc'] <= 100000), 1, 0)
df_inputs_prepr['annual_inc:100K-120K'] = np.where((df_inputs_prepr['annual_inc'] > 100000) & (df_inputs_prepr['annual_inc'] <= 120000), 1, 0)
df_inputs_prepr['annual_inc:120K-140K'] = np.where((df_inputs_prepr['annual_inc'] > 120000) & (df_inputs_prepr['annual_inc'] <= 140000), 1, 0)
df_inputs_prepr['annual_inc:>140K'] = np.where((df_inputs_prepr['annual_inc'] > 140000), 1, 0)

df_inputs_prepr['mths_since_last_delinq:Missing'] = np.where((df_inputs_prepr['mths_since_last_delinq'].isnull()), 1, 0)
df_inputs_prepr['mths_since_last_delinq:0-3'] = np.where((df_inputs_prepr['mths_since_last_delinq'] >= 0) & (df_inputs_prepr['mths_since_last_delinq'] < 4), 1, 0)
df_inputs_prepr['mths_since_last_delinq:4-30'] = np.where((df_inputs_prepr['mths_since_last_delinq'] >= 4) & (df_inputs_prepr['mths_since_last_delinq'] < 31), 1, 0)
df_inputs_prepr['mths_since_last_delinq:31-56'] = np.where((df_inputs_prepr['mths_since_last_delinq'] >= 31) & (df_inputs_prepr['mths_since_last_delinq'] < 57), 1, 0)
df_inputs_prepr['mths_since_last_delinq:>=57'] = np.where((df_inputs_prepr['mths_since_last_delinq'] >= 57), 1, 0)

df_inputs_prepr['dti:<=1.4'] = np.where((df_inputs_prepr['dti'] <= 1.4), 1, 0)
df_inputs_prepr['dti:1.4-3.5'] = np.where((df_inputs_prepr['dti'] > 1.4) & (df_inputs_prepr['dti'] <= 3.5), 1, 0)
df_inputs_prepr['dti:3.5-7.7'] = np.where((df_inputs_prepr['dti'] > 3.5) & (df_inputs_prepr['dti'] <= 7.7), 1, 0)
df_inputs_prepr['dti:7.7-10.5'] = np.where((df_inputs_prepr['dti'] > 7.7) & (df_inputs_prepr['dti'] <= 10.5), 1, 0)
df_inputs_prepr['dti:10.5-16.1'] = np.where((df_inputs_prepr['dti'] > 10.5) & (df_inputs_prepr['dti'] <= 16.1), 1, 0)
df_inputs_prepr['dti:16.1-20.3'] = np.where((df_inputs_prepr['dti'] > 16.1) & (df_inputs_prepr['dti'] <= 20.3), 1, 0)
df_inputs_prepr['dti:20.3-21.7'] = np.where((df_inputs_prepr['dti'] > 20.3) & (df_inputs_prepr['dti'] <= 21.7), 1, 0)
df_inputs_prepr['dti:21.7-22.4'] = np.where((df_inputs_prepr['dti'] > 21.7) & (df_inputs_prepr['dti'] <= 22.4), 1, 0)
df_inputs_prepr['dti:22.4-35'] = np.where((df_inputs_prepr['dti'] > 22.4) & (df_inputs_prepr['dti'] <= 35), 1, 0)
df_inputs_prepr['dti:>35'] = np.where((df_inputs_prepr['dti'] > 35), 1, 0)

df_inputs_prepr['mths_since_last_record:Missing'] = np.where((df_inputs_prepr['mths_since_last_record'].isnull()), 1, 0)
df_inputs_prepr['mths_since_last_record:0-2'] = np.where((df_inputs_prepr['mths_since_last_record'] >= 0) & (df_inputs_prepr['mths_since_last_record'] <= 2), 1, 0)
df_inputs_prepr['mths_since_last_record:3-20'] = np.where((df_inputs_prepr['mths_since_last_record'] >= 3) & (df_inputs_prepr['mths_since_last_record'] <= 20), 1, 0)
df_inputs_prepr['mths_since_last_record:21-31'] = np.where((df_inputs_prepr['mths_since_last_record'] >= 21) & (df_inputs_prepr['mths_since_last_record'] <= 31), 1, 0)
df_inputs_prepr['mths_since_last_record:32-80'] = np.where((df_inputs_prepr['mths_since_last_record'] >= 32) & (df_inputs_prepr['mths_since_last_record'] <= 80), 1, 0)
df_inputs_prepr['mths_since_last_record:81-86'] = np.where((df_inputs_prepr['mths_since_last_record'] >= 81) & (df_inputs_prepr['mths_since_last_record'] <= 86), 1, 0)
df_inputs_prepr['mths_since_last_record:>=86'] = np.where((df_inputs_prepr['mths_since_last_record'] >= 86), 1, 0)

### When running this check notebook 2 to replicate process

loan_data_inputs_2015 = df_inputs_prepr
loan_data_targets_2015 = df_targets_prepr

loan_data_inputs_2015.to_csv(r'D:\Datasets\LendingClub\loan_data_inputs_2015.csv')
loan_data_targets_2015.to_csv(r'D:\Datasets\LendingClub\loan_data_targets_2015.csv')
