==> 7-PD Model - Monitoring - PSI Calculations.ipynb <==
import numpy as np
import pandas as pd
pd.options.display.max_columns = None

inputs_train_with_ref_cat = pd.read_csv(r'D:\Datasets\LendingClub\inputs_train_with_ref_cat.csv', index_col=0)

loan_data_inputs_2015 = pd.read_csv(r'D:\Datasets\LendingClub\loan_data_inputs_2015.csv')
loan_data_targets_2015 = pd.read_csv(r'D:\Datasets\LendingClub\loan_data_targets_2015.csv')

loan_data_inputs_2015_with_ref_cat = loan_data_inputs_2015[inputs_train_with_ref_cat.columns]

loan_data_inputs_2015_with_ref_cat.shape

inputs_train_with_ref_cat.shape

df_scorecard = pd.read_csv(r'C:\Users\User\Desktop\Credit-Risk-Modelling-Python\df_scorecard.csv', index_col=0)
df_scorecard

inputs_train_with_ref_cat_w_intercept = inputs_train_with_ref_cat
inputs_train_with_ref_cat_w_intercept.insert(0, 'Intercept', 1)
cols = df_scorecard['Feature name'].values
cols = [c for c in cols if c not in ['grade:G', 'home_ownership:RENT_OTHER_NONE_ANY', 'addr_state:ND_NE_IA_NV_HI_FL_AL', 'verification_status:Verified', 'purpose:edu_sb_re_m_h_other', 'initial_list_status:f', 'term:60', 'emp_length:0', 'mths_since_issue_d:>84', 'int_rate:>20.281', 'mths_since_earliest_cr_line:<142', 'delinq_2yrs:>=4', 'inq_last_6mths:>6', 'open_acc:0', 'pub_rec:>=5', 'total_acc:<=25', 'acc_now_delinq:>=1', 'total_rev_hi_lim<=5K', 'annual_inc:<20K', 'mths_since_last_delinq:Missing', 'dti>35', 'mths_since_last_record:Missing']]
inputs_train_with_ref_cat_w_intercept = inputs_train_with_ref_cat_w_intercept[cols]
inputs_train_with_ref_cat_w_intercept.head()

inputs_2015_train_with_ref_cat_w_intercept = loan_data_inputs_2015_with_ref_cat
inputs_2015_train_with_ref_cat_w_intercept.insert(0, 'Intercept', 1)
cols = df_scorecard['Feature name'].values
cols = [c for c in cols if c not in ['grade:G', 'home_ownership:RENT_OTHER_NONE_ANY', 'addr_state:ND_NE_IA_NV_HI_FL_AL', 'verification_status:Verified', 'purpose:edu_sb_re_m_h_other', 'initial_list_status:f', 'term:60', 'emp_length:0', 'mths_since_issue_d:>84', 'int_rate:>20.281', 'mths_since_earliest_cr_line:<142', 'delinq_2yrs:>=4', 'inq_last_6mths:>6', 'open_acc:0', 'pub_rec:>=5', 'total_acc:<=25', 'acc_now_delinq:>=1', 'total_rev_hi_lim<=5K', 'annual_inc:<20K', 'mths_since_last_delinq:Missing', 'dti>35', 'mths_since_last_record:Missing']]
inputs_2015_train_with_ref_cat_w_intercept = inputs_2015_train_with_ref_cat_w_intercept[cols]
inputs_2015_train_with_ref_cat_w_intercept.head()

scorecard_scores = df_scorecard.copy()
scorecard_scores = scorecard_scores[scorecard_scores['Feature name'].isin(cols)]
scorecard_scores = scorecard_scores['Score - Final']
scorecard_scores = scorecard_scores.values.reshape(-1, 1)
scorecard_scores.shape

y_scores_train = inputs_train_with_ref_cat_w_intercept.dot(scorecard_scores)
y_scores_train.head()

y_scores_2015 = inputs_2015_train_with_ref_cat_w_intercept.dot(scorecard_scores)
y_scores_2015.head()

inputs_train_with_ref_cat_w_intercept = pd.concat([inputs_train_with_ref_cat_w_intercept, y_scores_train], axis=1)
inputs_2015_train_with_ref_cat_w_intercept = pd.concat([inputs_2015_train_with_ref_cat_w_intercept, y_scores_2015], axis=1)

inputs_train_with_ref_cat_w_intercept.columns.values[inputs_train_with_ref_cat_w_intercept.shape[1] - 1] = 'Score'
inputs_2015_train_with_ref_cat_w_intercept.columns.values[inputs_2015_train_with_ref_cat_w_intercept.shape[1] - 1] = 'Score'

inputs_2015_train_with_ref_cat_w_intercept.head()

# Dummy variables for score
inputs_train_with_ref_cat_w_intercept['Score:300-350'] = np.where((inputs_train_with_ref_cat_w_intercept['Score'] >= 300) & (inputs_train_with_ref_cat_w_intercept['Score'] < 350), 1, 0)
inputs_train_with_ref_cat_w_intercept['Score:350-400'] = np.where((inputs_train_with_ref_cat_w_intercept['Score'] >= 350) & (inputs_train_with_ref_cat_w_intercept['Score'] < 400), 1, 0)
inputs_train_with_ref_cat_w_intercept['Score:400-450'] = np.where((inputs_train_with_ref_cat_w_intercept['Score'] >= 400) & (inputs_train_with_ref_cat_w_intercept['Score'] < 450), 1, 0)
inputs_train_with_ref_cat_w_intercept['Score:450-500'] = np.where((inputs_train_with_ref_cat_w_intercept['Score'] >= 450) & (inputs_train_with_ref_cat_w_intercept['Score'] < 500), 1, 0)
inputs_train_with_ref_cat_w_intercept['Score:500-550'] = np.where((inputs_train_with_ref_cat_w_intercept['Score'] >= 500) & (inputs_train_with_ref_cat_w_intercept['Score'] < 550), 1, 0)
inputs_train_with_ref_cat_w_intercept['Score:550-600'] = np.where((inputs_train_with_ref_cat_w_intercept['Score'] >= 550) & (inputs_train_with_ref_cat_w_intercept['Score'] < 600), 1, 0)
inputs_train_with_ref_cat_w_intercept['Score:600-650'] = np.where((inputs_train_with_ref_cat_w_intercept['Score'] >= 600) & (inputs_train_with_ref_cat_w_intercept['Score'] < 650), 1, 0)
inputs_train_with_ref_cat_w_intercept['Score:650-700'] = np.where((inputs_train_with_ref_cat_w_intercept['Score'] >= 650) & (inputs_train_with_ref_cat_w_intercept['Score'] < 700), 1, 0)
inputs_train_with_ref_cat_w_intercept['Score:700-750'] = np.where((inputs_train_with_ref_cat_w_intercept['Score'] >= 700) & (inputs_train_with_ref_cat_w_intercept['Score'] < 750), 1, 0)
inputs_train_with_ref_cat_w_intercept['Score:750-800'] = np.where((inputs_train_with_ref_cat_w_intercept['Score'] >= 750) & (inputs_train_with_ref_cat_w_intercept['Score'] < 800), 1, 0)
inputs_train_with_ref_cat_w_intercept['Score:800-850'] = np.where((inputs_train_with_ref_cat_w_intercept['Score'] >= 800) & (inputs_train_with_ref_cat_w_intercept['Score'] < 850), 1, 0)

# Dummy variables for score
inputs_2015_train_with_ref_cat_w_intercept['Score:300-350'] = np.where((inputs_2015_train_with_ref_cat_w_intercept['Score'] >= 300) & (inputs_2015_train_with_ref_cat_w_intercept['Score'] < 350), 1, 0)
inputs_2015_train_with_ref_cat_w_intercept['Score:350-400'] = np.where((inputs_2015_train_with_ref_cat_w_intercept['Score'] >= 350) & (inputs_2015_train_with_ref_cat_w_intercept['Score'] < 400), 1, 0)
inputs_2015_train_with_ref_cat_w_intercept['Score:400-450'] = np.where((inputs_2015_train_with_ref_cat_w_intercept['Score'] >= 400) & (inputs_2015_train_with_ref_cat_w_intercept['Score'] < 450), 1, 0)
inputs_2015_train_with_ref_cat_w_intercept['Score:450-500'] = np.where((inputs_2015_train_with_ref_cat_w_intercept['Score'] >= 450) & (inputs_2015_train_with_ref_cat_w_intercept['Score'] < 500), 1, 0)
inputs_2015_train_with_ref_cat_w_intercept['Score:500-550'] = np.where((inputs_2015_train_with_ref_cat_w_intercept['Score'] >= 500) & (inputs_2015_train_with_ref_cat_w_intercept['Score'] < 550), 1, 0)
inputs_2015_train_with_ref_cat_w_intercept['Score:550-600'] = np.where((inputs_2015_train_with_ref_cat_w_intercept['Score'] >= 550) & (inputs_2015_train_with_ref_cat_w_intercept['Score'] < 600), 1, 0)
inputs_2015_train_with_ref_cat_w_intercept['Score:600-650'] = np.where((inputs_2015_train_with_ref_cat_w_intercept['Score'] >= 600) & (inputs_2015_train_with_ref_cat_w_intercept['Score'] < 650), 1, 0)
inputs_2015_train_with_ref_cat_w_intercept['Score:650-700'] = np.where((inputs_2015_train_with_ref_cat_w_intercept['Score'] >= 650) & (inputs_2015_train_with_ref_cat_w_intercept['Score'] < 700), 1, 0)
inputs_2015_train_with_ref_cat_w_intercept['Score:700-750'] = np.where((inputs_2015_train_with_ref_cat_w_intercept['Score'] >= 700) & (inputs_2015_train_with_ref_cat_w_intercept['Score'] < 750), 1, 0)
inputs_2015_train_with_ref_cat_w_intercept['Score:750-800'] = np.where((inputs_2015_train_with_ref_cat_w_intercept['Score'] >= 750) & (inputs_2015_train_with_ref_cat_w_intercept['Score'] < 800), 1, 0)
inputs_2015_train_with_ref_cat_w_intercept['Score:800-850'] = np.where((inputs_2015_train_with_ref_cat_w_intercept['Score'] >= 800) & (inputs_2015_train_with_ref_cat_w_intercept['Score'] < 850), 1, 0)

## Calculating PSI (population stability index) for features

PSI_calc_train = inputs_train_with_ref_cat_w_intercept.sum() / inputs_train_with_ref_cat_w_intercept.shape[0]

PSI_calc_2015 = inputs_2015_train_with_ref_cat_w_intercept.sum() / inputs_2015_train_with_ref_cat_w_intercept.shape[0]

PSI_calc = pd.concat([PSI_calc_train, PSI_calc_2015], axis=1)
PSI_calc

PSI_calc = PSI_calc.reset_index()
PSI_calc['Original feature name'] = PSI_calc['index'].str.split(':').str[0]
PSI_calc.columns = ['index', 'Proportions_Train', 'Proportions_New', 'Original feature name']

PSI_calc = PSI_calc[np.array(['index', 'Original feature name', 'Proportions_Train', 'Proportions_New'])]

PSI_calc

PSI_calc = PSI_calc[(PSI_calc['index'] != 'Intercept') & (PSI_calc['index'] != 'Score')]

# PSI = sigma[(%Actualj - %Expectedj) * ln(%Actualj/%Expectedj)]

PSI_calc['Contribution'] = np.where((PSI_calc['Proportions_Train'] == 0) | (PSI_calc['Proportions_New'] == 0), 0,
                                   (PSI_calc['Proportions_New'] - PSI_calc['Proportions_Train']) * np.log(PSI_calc['Proportions_New']/PSI_calc['Proportions_Train']))

PSI_calc

PSI_calc.groupby('Original feature name')['Contribution'].sum()

PSI_calc[PSI_calc['Original feature name']=='mths_since_issue_d']

# PSI
# 0 and 1
# 0 = no difference
# 1 = absolute difference
# < 0.1 = little to no difference
# 0.1 < PSI < 0.25 = little difference (no action taken)
# PSI > 0.25 = big difference (action taken)
